+++ lib/ecto_shorts/query_builder.ex
@@ -1,8 +1,6 @@
 defmodule EctoShorts.QueryBuilder do
   @moduledoc false
 
-  alias EctoShorts.QueryBuilder.Postgres
-
   @callback build_query(
               source :: any(),
               query :: any(),
@@ -11,12 +9,4 @@ defmodule EctoShorts.QueryBuilder do
               args :: any(),
               opts :: keyword()
             ) :: Ecto.Query.t()
-
-  def build_query(source, query, binding_selector, current_filter, args, opts \\ []) do
-    adapter(opts).build_query(source, query, binding_selector, current_filter, args, opts)
-  end
-
-  defp adapter(opts) do
-    opts[:query_builder] || Postgres
-  end
 end