+++ lib/ecto_shorts/query_builders/postgres/dynamics/common_expr.ex
@@ -1,0 +1,10 @@
+defmodule EctoShorts.QueryBuilders.Postgres.Dynamics.CommonExpr do
+  @moduledoc false
+
+  alias EctoShorts.QueryBuilders.Postgres.Dynamics.CommonExprBuilder
+
+  require Ecto.Query
+  require CommonExprBuilder
+
+  CommonExprBuilder.define_common_exprs()
+end