+++ lib/ecto_shorts/query_builders/postgres/dynamics/scalar_expr.ex
@@ -1,0 +1,10 @@
+defmodule EctoShorts.QueryBuilders.Postgres.Dynamics.ScalarExpr do
+  @moduledoc false
+
+  alias EctoShorts.QueryBuilders.Postgres.Dynamics.ScalarExprBuilder
+
+  require Ecto.Query
+  require ScalarExprBuilder
+
+  ScalarExprBuilder.define_scalar_exprs()
+end